chore(coin-tron): outsource wallet types
