+++ .changeset/blue-toys-add.md
@@ -1,0 +1,7 @@
+---
+"@ledgerhq/coin-tester-tron": patch
+"@ledgerhq/coin-tron": major
+"@ledgerhq/live-common": patch
+---
+
+chore(coin-tron): outsource wallet types
