+++ libs/coin-modules/coin-tron/src/types/index.ts
@@ -1,4 +1,3 @@
-export * from "./bridge";
+export * from "./model";
 export * from "./errors";
-export * from "./signer";
 export * from "./assets";
