--- libs/coin-modules/coin-tron/src/types/signer.ts
@@ -1,9 +1,0 @@
-export type TronAddress = {
-  publicKey: string;
-  address: string;
-};
-export type TronSignature = string;
-export interface TronSigner {
-  getAddress(path: string, boolDisplay?: boolean): Promise<TronAddress>;
-  sign(path: string, rawTxHex: string, tokenSignatures: string[]): Promise<TronSignature>;
-}
