+++ libs/coin-tester-modules/coin-tester-tron/src/scenarii/tron.ts
@@ -14,8 +14,8 @@ import {
   makeTrc20Token,
   registerTronTokensInMockStore,
 } from "../fixtures";
-import type { TronAccount } from "@ledgerhq/coin-tron/types/index";
 import type {
+  TronAccount,
   Transaction as TronTransaction,
   TronFamilySpecificData,
 } from "@ledgerhq/live-common/families/tron/types";
