+++ libs/ledger-live-common/src/families/tron/data.mock.ts
@@ -1,5 +1,5 @@
 import { fromAccountRaw } from "../../account/serialization";
-import { TronAccount } from "@ledgerhq/coin-tron/types/index";
+import type { TronAccount } from "./types";
 import { setCryptoAssetsStore } from "@ledgerhq/ledger-wallet-framework/cryptoAssetsStore";
 
 export const __NEXT_REWARD_DATE__ = new Date(Date.now() - 6 * 60 * 60 * 1000);
