+++ libs/ledger-live-common/src/mock/account.ts
@@ -14,7 +14,7 @@ import { CosmosAccount } from "../families/cosmos/types";
 import { BitcoinAccount } from "@ledgerhq/coin-bitcoin/types";
 import { PolkadotAccount } from "@ledgerhq/coin-polkadot/types/index";
 import type { TezosAccount } from "../families/tezos/types";
-import { TronAccount } from "@ledgerhq/coin-tron/types/index";
+import type { TronAccount } from "../families/tron/types";
 import { CardanoAccount, PaymentChain } from "@ledgerhq/coin-cardano/types";
 import { types } from "@stricahq/typhonjs";
 import { SolanaAccount } from "@ledgerhq/coin-solana/types";
